Move view to another file
